# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sorveteria API",
        "version": "1.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login efetuado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "JWT token",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credendiais inválidas"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Logout",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Desconectado com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/sabores": {
            "get": {
                "tags": [
                    "Sabores"
                ],
                "summary": "Listar todos os sabores",
                "operationId": "27d3839232bf9305e325bb564139956f",
                "responses": {
                    "200": {
                        "description": "Lista de sabores"
                    }
                }
            }
        },
        "/api/sabores/{id}": {
            "get": {
                "tags": [
                    "Sabores"
                ],
                "summary": "Buscar sabor por ID",
                "operationId": "1e26946c7c3d671cfc41b56e283b9102",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do sabor",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do sabor"
                    },
                    "404": {
                        "description": "Sabor não encontrado"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Login",
            "description": "Login"
        },
        {
            "name": "Sabores",
            "description": "Sabores"
        }
    ],
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "description": "JWT Authorization header using the Bearer scheme."
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}